+++ .formatter.exs
@@ -1,10 +1,4 @@
-locals_without_parens = [
-  before_build_policy: 1
-]
-
 [
   import_deps: [:ecto, :ecto_sql],
-  inputs: ["{mix,.formatter}.exs", "{config,lib,test}/**/*.{ex,exs}"],
-  locals_without_parens: locals_without_parens,
-  export: [locals_without_parens: locals_without_parens]
+  inputs: ["{mix,.formatter}.exs", "{config,lib,test}/**/*.{ex,exs}"]
 ]