+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
   * [Janus] `use Janus` accepts optional defaults for `Janus.Authorization.authorize/4` options
   * [Janus.Policy] **Breaking:** `policy_for` renamed to `build_policy`
+  * [Janus.Policy] **Breaking:** `before_policy_for` removed
   * [Janus.Authorization] **Breaking:** `filter_authorized` renamed to `scope`
   * [Janus.Authorization] **Breaking:** `authorize` now returns `{:error, :not_authorized}` instead of `:error`
   * [Janus.Authorization] `authorize/4` adds `:repo` and `:load_assocations` options